[UPD] Include simple support for ignoring hooks flags
This allows disabling and re-enabling of the `forbid-git-conflicts` and
`forbid-set-trace` hooks, using comments to turn it off and on, similar
to `black`'s `#fmt: off` and `#fmt: on`.

```
# forbid-git-conflicts: off
"<<<<<<<"
# forbid-git-conflicts: on
```

```
# forbid-set-trace: off
__import__("ipdb").set_trace()
# forbid-set-trace: on
```

These options can be useful for ignoring "bad content" that is being
flagged as a false positive, such as comments, documentation, etc.